+++ climaf/cmacros.py
@@ -10,7 +10,7 @@ CliMAF macros module :
 
 import sys, os
 
-from classes import cobject, cdataset, ctree, scriptChild
+from classes import cobject, cdataset, ctree, scriptChild, cpage
 from clogging import clogger, dedent
 
 #: Dictionnary of macros
@@ -28,7 +28,7 @@ class cdummy(cobject):
     def buildcrs(self,period=None,crsrewrite=None):
         return('ARG')
 
-def cmacro(name,cobject,lobjects=[]):
+def cmacro(name,cobj,lobjects=[]):
     """
 
     Define a CliMAF macro from a CliMAF compound object.
@@ -40,7 +40,7 @@ def cmacro(name,cobject,lobjects=[]):
     Args:
      name (string) : the name you want to give to the macro; a Python
       function with the same name will be defined
-     cobject (CliMAF object, or string) : any CliMAF object, usually
+     cobj (CliMAF object, or string) : any CliMAF object, usually
       the result of a series of operators, that you would like to
       repeat using other input datasets; alternatively, you can provide
       the macro formula as a string (when accustomed to the syntax)
@@ -52,56 +52,62 @@ def cmacro(name,cobject,lobjects=[]):
       a macro; the returned value is usualy not used asis; a python
       function is also defined in module cmacros, and you may import it
       and use it in the same way as a CliMAF operator. All the datasets
-      involved in ``cobject`` become arguments of the macro, which
+      involved in ``cobj`` become arguments of the macro, which
       allows you to re-do the same computations and easily define
-      objects similar to ``cobjects``
+      objects similar to ``cobjs``
 
     Example::
 
-     # First use and combine CliMAF operators to get some interesting result using some dataset(s)
-     january_ta=ds(project='example',experiment='AMIPV6ALB2G',variable='ta',frequency='monthly',period='198001')
-     ta_europe=llbox(january_ta,latmin=40,latmax=60,lonmin=-15,lonmax=25)
-     ta_ezm=ccdo(ta_europe,operator='zonmean')
-     fig_ezm=plotxsec(ta_ezm)
-     #
-     # Using this result as an example, define a macro named 'eu_cross_section',
-     # which arguments will be the datasets involved in this result
-     cmacro('eu_cross_section',fig_ezm)
-     #
-     # You can of course apply a macro to other dataset(s), 
-     pr=ds(project='example',experiment='AMIPV6ALB2G', variable='pr', frequency='monthly', period='198001')
-     pr_ezm=climaf.cmacros.eu_zonal_mean(pr)
-     #
-     # A simpler way to use it ; first import the macro
-     from climaf.cmacros import eu_zonal_mean
-     # or :
-     from climaf.cmacros import *
-     pr_ezm=eu_zonal_mean(pr)
-     #
-     # All macros are registered in dictionnary climaf.cmacros.cmacros,
-     # which is imported by climaf.api; you can list it by :
-     cmacros
+     >>> # First use and combine CliMAF operators to get some interesting result using some dataset(s)
+     >>> january_ta=ds(project='example',experiment='AMIPV6ALB2G',variable='ta',frequency='monthly',period='198001')
+     >>> ta_europe=llbox(january_ta,latmin=40,latmax=60,lonmin=-15,lonmax=25)
+     >>> ta_ezm=ccdo(ta_europe,operator='zonmean')
+     >>> fig_ezm=plotxsec(ta_ezm)
+     >>> #
+     >>> # Using this result as an example, define a macro named 'eu_cross_section',
+     >>> # which arguments will be the datasets involved in this result
+     >>> cmacro('eu_cross_section',fig_ezm)
+     >>> #
+     >>> # You can of course apply a macro to other dataset(s), 
+     >>> pr=ds(project='example',experiment='AMIPV6ALB2G', variable='pr', frequency='monthly', period='198001')
+     >>> pr_ezm=climaf.cmacros.eu_zonal_mean(pr)
+     >>> #
+     >>> # A simpler way to use it ; first import the macro
+     >>> from climaf.cmacros import eu_zonal_mean
+     >>> # or :
+     >>> from climaf.cmacros import *
+     >>> pr_ezm=eu_zonal_mean(pr)
+     >>> #
+     >>> # All macros are registered in dictionnary climaf.cmacros.cmacros,
+     >>> # which is imported by climaf.api; you can list it by :
+     >>> cmacros
 
 
     See also much more explanations in the example at :download:`macro.py <../examples/macro.py>`
     """
-    if isinstance(cobject,str) :
+    if isinstance(cobj,str) :
         # Next line used for interpreting macros's CRS
         exec("ARG=climaf.cmacros.cdummy()", sys.modules['__main__'].__dict__)
-        cobject=eval(cobject, sys.modules['__main__'].__dict__)
+        cobj=eval(cobj, sys.modules['__main__'].__dict__)
     domatch=False
     for o in lobjects :
-        domatch = domatch or cobject.buildcrs() == o.buildcrs() 
-    if isinstance(cobject,cdataset) or isinstance(cobject,cdummy) or domatch :
+        domatch = domatch or cobj==o or \
+                  ( isinstance(cobj,cobject) and cobj.buildcrs() == o.buildcrs())
+    if isinstance(cobj,cdataset) or isinstance(cobj,cdummy) or domatch :
         return cdummy()
-    elif isinstance(cobject,ctree) :
-        rep=ctree(cobject.operator, cobject.script,
-                  *cobject.operands, **cobject.parameters)
+    elif isinstance(cobj,ctree) :
+        rep=ctree(cobj.operator, cobj.script,
+                  *cobj.operands, **cobj.parameters)
         rep.operands = map(cmacro,[ None for o in rep.operands],rep.operands)
-    elif isinstance(cobject,scriptChild) :
-        rep=scriptChild(cmacro(None,cobject.father),cobject.varname)
+    elif isinstance(cobj,scriptChild) :
+        rep=scriptChild(cmacro(None,cobj.father),cobj.varname)
+    elif isinstance(cobj,cpage) :
+        rep=cpage(cobj.widths_list, cobj.heights_list,
+                  [ map(cmacro, [ None for fig in line ], line) for line in cobj.fig_lines ] ,
+                  cobj.orientation)
+    elif cobj is None : return None
     else :
-        clogger.error("Cannot yet handle object :%s", `cobject`)
+        clogger.error("Cannot yet handle object :%s", `cobj`)
         rep=None
     if name and rep :
         cmacros[name]=rep
@@ -127,11 +133,11 @@ def crewrite(crs,alsoAtTop=True):
     exec("ARG=climaf.cmacros.cdummy()", sys.modules['__main__'].__dict__)
     #
     co=eval(crs, sys.modules['__main__'].__dict__)
-    if isinstance(co,ctree) or isinstance(co,scriptChild) :
+    if isinstance(co,ctree) or isinstance(co,scriptChild) or isinstance(co,cpage) :
         if alsoAtTop :
             for m in cmacros :
                 clogger.debug("looking at macro : "+m+"="+`cmacros[m]`+\
-                           " compared to : "+`cmacro(None,co)`)
+                           " \ncompared to : "+`cmacro(None,co)`)
                 argl=cmatch(cmacros[m],co)
                 if len(argl) > 0 :
                     rep=m+"("
@@ -145,48 +151,52 @@ def crewrite(crs,alsoAtTop=True):
     else :
         return(crs)
 
-def cmatch(macro,cobject) :
+def cmatch(macro,cobj) :
     """
-    Analyze if macro does match cobject, and return the list of objects
+    Analyze if macro does match cobj, and return the list of objects
     matching macro arguments, ordered by depth-first traversal
     """
-    clogger.debug("matching "+`macro`+" and "+`cobject`)
-    if isinstance(cobject,ctree) and isinstance(macro,ctree) and \
-           macro.operator==cobject.operator :
+    clogger.debug("matching "+`macro`+" and "+`cobj`)
+    if isinstance(cobj,ctree) and isinstance(macro,ctree) and \
+           macro.operator==cobj.operator :
         nok=False
-        for mpara,para in zip(macro.parameters,cobject.parameters) :
+        for mpara,para in zip(macro.parameters,cobj.parameters) :
             if mpara != para or \
-                   macro.parameters[para] != cobject.parameters[para]:
+                   macro.parameters[para] != cobj.parameters[para]:
                 nok=True
         if nok : return []
         argsub=[]
-        for mop,op in zip(macro.operands,cobject.operands) :
+        for mop,op in zip(macro.operands,cobj.operands) :
             if isinstance(mop,cdummy) :
                 argsub.append(op)
             else :
                 argsub+=cmatch(mop,op)
         return(argsub)
-    elif isinstance(cobject,scriptChild) and isinstance(macro,scriptChild) and \
-           macro.varname==cobject.varname :
-        return(cmatch(macro.father,cobject.father,argslist))
+    elif isinstance(cobj,scriptChild) and isinstance(macro,scriptChild) and \
+           macro.varname==cobj.varname :
+        return(cmatch(macro.father,cobj.father,argslist))
     else : return []
     
     
 def read(filename=macroFilename):
     """
-    Read macro dictionnary in filenmae, and add it to cmacros[]
+    Read macro dictionnary from filename, and add it to cmacros[]
     """
     import json
     global cmacros
     macros_texts=None
     try :
         macrofile=file(os.path.expanduser(filename), "r")
+        clogger.debug("Macrofile %s read"%(macrofile))
         macros_texts=json.load(macrofile)
+        clogger.debug("After reading file %s, macros=%s"%(macrofile,`macros_texts`))
         macrofile.close()
     except:
         clogger.info("Issue reading macro file %s ", filename)
     if macros_texts :
-        for m in macros_texts : cmacro(macros_texts[m],m)
+        for m in macros_texts :
+            clogger.debug("loading macro %s"%m)
+            cmacro(m,macros_texts[m])
 
 
 def write(filename=macroFilename) :