+++ climaf/operators.py
@@ -8,7 +8,7 @@ CliMAF handling of external scripts and binaries , and of internal operators (Py
 
 import os, re, sys, subprocess
 import driver
-from clogging import clogger
+from clogging import clogger, dedent
 
 # Next definition can be splitted in a set managed by an administrator, and
 # other sets managed and fed by users. But it should be enforced that no redefinition
@@ -413,8 +413,4 @@ if __name__ == "__main__":
     cscript('test_script' ,'echo $*')
     test_script(arg1=1,arg2='two')
 
-#scripts['eof']='eof "${in}" ${out_eof} ${out_cp}'
-#scripts['regress']='regress ${in_1} ${in_2} ${out_regress} ${out_correl}'
-#scripts['spatial_average']="cdo fldavg ${datain} ${dataout} ${box} ${dataout2}"
-#scripts['eof']="eof ${datain} ${out1} ${dataout2}"